Optimisation du formulaire d'authentification
